backend/ml_service/requirements.txt
numpy==1.21.0
pandas==1.3.0
scikit-learn==0.24.2
flask==2.0.1
pymongo==3.12.0
python-dotenv==0.19.0
lightfm==1.16
scipy==1.7.0
requests==2.26.0 